Gold-YOLO_jittor/requirements.txt
# Gold-YOLO Jittor Implementation Requirements
# Python >= 3.7 environment

jittor>=1.3.0
numpy>=1.18.5
opencv-python>=4.1.2
PyYAML>=5.3.1
scipy>=1.4.1
tqdm>=4.41.0
pillow>=8.0.0

# Optional for visualization and evaluation
matplotlib>=3.3.0
seaborn>=0.11.0

# For data processing
pycocotools>=2.0
